Updates using .exits() to validate queryset without evaluating combine management commands into one
